+++ development/image-builder/pkg/builder/dockerfile.go
@@ -1,0 +1,9 @@
+package builder
+
+import (
+	"github.com/kyma-project/test-infra/development/image-builder/pkg/config"
+)
+
+type Builder struct {
+	config.Image
+}
