--- image-builder/pkg/builder/dockerfile.go
@@ -1,7 +1,0 @@
-package builder
-
-import "github.com/kyma-project/test-infra/image-builder/pkg/config"
-
-type Builder struct {
-	config.Image
-}
