+++ prow/jobs/test-infra/image-builder.yaml
@@ -1,0 +1,43 @@
+# Code generated by rendertemplates. DO NOT EDIT.
+
+
+presubmits: # runs on PRs
+  kyma-project/test-infra:
+    - name: pre-test-infra-build-testimages
+      annotations:
+        pipeline.trigger: "pr-submit"
+      labels:
+        prow.k8s.io/pubsub.project: "sap-kyma-prow"
+        prow.k8s.io/pubsub.runID: "pre-test-infra-build-testimages"
+        prow.k8s.io/pubsub.topic: "prowjobs"
+      run_if_changed: '^prow/images/'
+      skip_report: false
+      decorate: true
+      cluster: untrusted-workload
+      max_concurrency: 10
+      branches:
+        - ^master$
+        - ^main$
+      spec:
+        containers:
+          - image: "eu.gcr.io/sap-kyma-neighbors-dev/image-builder:test"
+            securityContext:
+              privileged: true
+            command:
+              - "/image-builder.sh"
+            args:
+              - "-c"
+              - "prow/images/build-config.yaml"
+              - "--debug"
+            resources:
+              requests:
+                memory: 3Gi
+                cpu: 2
+        tolerations:
+          - key: dedicated
+            operator: Equal
+            value: high-cpu
+            effect: NoSchedule
+        nodeSelector:
+            dedicated: "high-cpu"
+
